+++ test/conf/new/protractor.saucelabs.conf.js
@@ -4,24 +4,31 @@ let config = require('../protractor.shared.conf.js').config;
 const basicSpecs = '../../basics.spec.js';
 const deskSpecs = '../../new.desktop.spec.js';
 const mobileSpecs = '../../new.mobile.spec.js';
+const screenResolution = '1600x1200';
+const shardTestFiles = true;
+const tunnelIdentifier = process.env.TRAVIS_JOB_NUMBER;
+const defaultCapabilities = {
+	shardTestFiles: true,
+	tunnelIdentifier: process.env.TRAVIS_JOB_NUMBER,
+};
 
 config.sauceUser = process.env.IC_SAUCE_USERNAME ? process.env.IC_SAUCE_USERNAME : process.env.SAUCE_USERNAME;
 config.sauceKey = process.env.IC_SAUCE_ACCESS_KEY ? process.env.IC_SAUCE_ACCESS_KEY : process.env.SAUCE_ACCESS_KEY;
-config.sauceBuild = process.env.TRAVIS_JOB_NUMBER;
+config.sauceBuild = process.env.TRAVIS_JOB_NUMBER || 'local-build';
 
 config.multiCapabilities = [
 	/**
 	 * iOS
 	 */
+	// @todo: Tablets can have multiple tabs open, when this happens the address bar is higher and making selecting an element fail
 	{
 		deviceName: 'iPad Pro (12.9 inch) (2nd generation) Simulator',
 		browserName: 'safari',
 		logName: 'iPadPro12.9.2nd',
 		platformName: 'ios',
 		platformVersion: '12.0',
-		tunnelIdentifier: process.env.TRAVIS_JOB_NUMBER,
-		shardTestFiles: true,
-		specs: [ mobileSpecs ]
+		specs: [ mobileSpecs ],
+		...defaultCapabilities,
 	},
 	{
 		deviceName: 'iPad Pro (9.7 inch) Simulator',
@@ -29,9 +36,8 @@ config.multiCapabilities = [
 		logName: 'iPadPro9.7Simulator',
 		platformName: 'ios',
 		platformVersion: '11.3',
-		tunnelIdentifier: process.env.TRAVIS_JOB_NUMBER,
-		shardTestFiles: true,
-		specs: [ mobileSpecs ]
+		specs: [ mobileSpecs ],
+		...defaultCapabilities,
 	},
 	{
 		deviceName: 'iPad Air Simulator',
@@ -39,9 +45,8 @@ config.multiCapabilities = [
 		logName: 'iPadAirSimulator',
 		platformName: 'ios',
 		platformVersion: '10.3',
-		tunnelIdentifier: process.env.TRAVIS_JOB_NUMBER,
-		shardTestFiles: true,
-		specs: [ mobileSpecs ]
+		specs: [ mobileSpecs ],
+		...defaultCapabilities,
 	},
 	{
 		browserName: 'safari',
@@ -49,19 +54,18 @@ config.multiCapabilities = [
 		logName: 'iPhone8Simulator',
 		platformName: 'ios',
 		platformVersion: '12.0',
-		tunnelIdentifier: process.env.TRAVIS_JOB_NUMBER,
-		shardTestFiles: true,
-		specs: [ mobileSpecs ]
+		specs: [ mobileSpecs ],
+		...defaultCapabilities,
 	},
+	// @TODO: check iPhone X series toolbar and black thingy
 	{
 		browserName: 'safari',
 		deviceName: 'iPhone X Simulator',
 		logName: 'iPhoneXSimulator',
 		platformName: 'ios',
 		platformVersion: '12.0',
-		tunnelIdentifier: process.env.TRAVIS_JOB_NUMBER,
-		shardTestFiles: true,
-		specs: [ mobileSpecs ]
+		specs: [ mobileSpecs ],
+		...defaultCapabilities,
 	},
 	// Not supported in the cloud now
 	// {
@@ -81,10 +85,9 @@ config.multiCapabilities = [
 		logName: 'GooglePixelGoogleAPIEmulator7.1NativeWebScreenshot',
 		platformName: 'Android',
 		platformVersion: '7.1',
-		tunnelIdentifier: process.env.TRAVIS_JOB_NUMBER,
-		shardTestFiles: true,
 		specs: [ mobileSpecs ],
 		nativeWebScreenshot: true,
+		...defaultCapabilities,
 	},
 	{
 		browserName: 'chrome',
@@ -92,10 +95,9 @@ config.multiCapabilities = [
 		logName: 'AndroidGoogleApiEmulator6.0NativeWebScreenshot',
 		platformName: 'Android',
 		platformVersion: '6.0',
-		tunnelIdentifier: process.env.TRAVIS_JOB_NUMBER,
-		shardTestFiles: true,
 		specs: [ mobileSpecs ],
 		nativeWebScreenshot: true,
+		...defaultCapabilities,
 	},
 	// Not supporting Android Tablets now with nativeWebScreenshot
 	// {
@@ -104,8 +106,8 @@ config.multiCapabilities = [
 	// 	logName: 'GooglePixelCTablet7.1NativeWebScreenshot',
 	// 	platformName: 'Android',
 	// 	platformVersion: '7.1',
-	// 	tunnelIdentifier: process.env.TRAVIS_JOB_NUMBER,
-	// 	shardTestFiles: true,
+	// 	tunnelIdentifier,
+	// 	shardTestFiles,
 	// 	specs: [ mobileSpecs ],
 	// 	nativeWebScreenshot: true,
 	// },
@@ -119,9 +121,8 @@ config.multiCapabilities = [
 		logName: 'GooglePixelGoogleAPIEmulator7.1ChromeDriver',
 		platformName: 'Android',
 		platformVersion: '7.1',
-		tunnelIdentifier: process.env.TRAVIS_JOB_NUMBER,
-		shardTestFiles: true,
 		specs: [ mobileSpecs ],
+		...defaultCapabilities,
 	},
 	{
 		browserName: 'chrome',
@@ -129,9 +130,7 @@ config.multiCapabilities = [
 		logName: 'AndroidGoogleApiEmulator6.0ChromeDriver',
 		platformName: 'Android',
 		platformVersion: '6.0',
-		tunnelIdentifier: process.env.TRAVIS_JOB_NUMBER,
-		shardTestFiles: true,
-		specs: [ mobileSpecs ],
+		specs: [ mobileSpecs ], ...defaultCapabilities,
 	},
 	{
 		browserName: 'chrome',
@@ -139,9 +138,8 @@ config.multiCapabilities = [
 		logName: 'GooglePixelCTablet7.1ChromeDriver',
 		platformName: 'Android',
 		platformVersion: '7.1',
-		tunnelIdentifier: process.env.TRAVIS_JOB_NUMBER,
-		shardTestFiles: true,
 		specs: [ mobileSpecs ],
+		...defaultCapabilities,
 	},
 
 	/**
@@ -151,81 +149,58 @@ config.multiCapabilities = [
 		browserName: 'chrome',
 		platform: 'Windows 10',
 		version: 'latest',
-		screenResolution: '1400x1050',
-		public: 'public',
-		tunnelIdentifier: process.env.TRAVIS_JOB_NUMBER,
 		logName: 'chrome-latest',
-		shardTestFiles: true,
 		specs: [ basicSpecs, deskSpecs ],
+		screenResolution,
+		...defaultCapabilities,
 	},
 	{
 		browserName: 'firefox',
 		platform: 'Windows 10',
 		version: 'latest',
-		screenResolution: '1400x1050',
-		public: 'public',
-		'tunnel-identifier': process.env.TRAVIS_JOB_NUMBER,
 		logName: 'Firefox latest',
-		shardTestFiles: true,
-		specs: [ deskSpecs ]
+		specs: [ deskSpecs ],
+		screenResolution,
+		...defaultCapabilities,
 	},
-
-	// {
-	//     browserName: 'firefox',
-	//     platform: "Windows 10",
-	//     version: "47",
-	//     screenResolution: "1400x1050",
-	//     public: "public",
-	//     "tunnel-identifier": process.env.TRAVIS_JOB_NUMBER,
-	//     logName: "Firefox 47",
-	//     shardTestFiles: true,
-	//     specs: [deskSpecs]
-	// },
-	// {
-	//     browserName: 'internet explorer',
-	//     platform: "Windows 8.1",
-	//     version: "11.0",
-	//     screenResolution: "1400x1050",
-	//     public: "public",
-	//     "tunnel-identifier": process.env.TRAVIS_JOB_NUMBER,
-	//     logName: "IE11",
-	//     shardTestFiles: true,
-	//     specs: [deskSpecs]
-	// },
-	// {
-	//     browserName: 'MicrosoftEdge',
-	//     platform: "Windows 10",
-	//     version: "latest",
-	//     screenResolution: "1400x1050",
-	//     public: "public",
-	//     "tunnel-identifier": process.env.TRAVIS_JOB_NUMBER,
-	//     logName: "Microsoft Edge latest",
-	//     shardTestFiles: true,
-	//     specs: [deskSpecs]
-	// },
-	// // Use 9 and 10 because of the different webdriver, 9 has an old and 10 a new
-	// {
-	//     browserName: 'safari',
-	//     platform: "OS X 10.11",
-	//     version: "9",
-	//     screenResolution: "1600x1200",
-	//     public: "public",
-	//     "tunnel-identifier": process.env.TRAVIS_JOB_NUMBER,
-	//     logName: "Safari 9",
-	//     shardTestFiles: true,
-	//     specs: [deskSpecs]
-	// },
-	// {
-	//     browserName: 'safari',
-	//     platform: "OS X 10.11",
-	//     version: "10",
-	//     screenResolution: "1600x1200",
-	//     public: "public",
-	//     "tunnel-identifier": process.env.TRAVIS_JOB_NUMBER,
-	//     logName: "Safari 10",
-	//     shardTestFiles: true,
-	//     specs: [deskSpecs]
-	// }
+	// @todo: Check why the cutout of IE11 is not properly made, looks like a few pixels of on the left
+	{
+		browserName: 'internet explorer',
+		platform: 'Windows 8.1',
+		version: 'latest',
+		logName: 'IE11',
+		specs: [ deskSpecs ],
+		screenResolution,
+		...defaultCapabilities,
+	},
+	// @todo: Currently having an issue selecting an element
+	{
+		browserName: 'MicrosoftEdge',
+		platform: 'Windows 10',
+		version: 'latest',
+		logName: 'Microsoft Edge latest',
+		specs: [ deskSpecs ],
+		screenResolution,
+		...defaultCapabilities,
+	},
+	{
+		browserName: 'safari',
+		platform: 'macOS 10.12',
+		version: '11.0',
+		logName: 'SierraSafari11',
+		specs: [ deskSpecs ],
+		screenResolution,
+		...defaultCapabilities,
+	},
+	{
+		browserName: 'safari',
+		platform: 'macOS 10.13',
+		version: '12.0',
+		logName: 'HiSierraSafari12',
+		specs: [ deskSpecs ],
+		screenResolution,
+		...defaultCapabilities,
+	}
 ];
 
 config.plugins = [ {