+++ test/new.desktop.spec.js
@@ -1,5 +1,8 @@
+/**
+ * These tests need to be changed to `check` tests in the future, for know the saveElement is sufficient
+ */
 describe('image compare basics', () => {
-  const headerElement = $('h1.uk-heading-large');
+	const firstButton = $('.uk-button:nth-child(1)');
 
   beforeEach(async () => {
     await browser.get(browser.baseUrl);
@@ -9,11 +12,15 @@ describe('image compare basics', () => {
 	// Chrome remembers the last postion when the url is loaded again, this will reset it.
 	afterEach(() => browser.executeScript('window.scrollTo(0, 0);'));
 
-  it('should do a save element with custom dimensions the deprecated way', async () => {
-    await browser.imageCompare.saveElement(headerElement, 'resizeDimensions-examplePageElement-deprecated', {resizeDimensions: 15});
+	it('should do a save element', async () => {
+		await browser.imageCompare.saveElement(firstButton, 'firstButtonElement', { empty: null });
+	});
+
+  xit('should do a save element with custom dimensions the deprecated way', async () => {
+    await browser.imageCompare.saveElement(firstButton, 'resizeDimensions-firstButtonElement-deprecated', {resizeDimensions: 15});
   });
 
-  it('should do a save element with custom dimensions the new way', async () => {
-    await browser.imageCompare.saveElement(headerElement, 'resizeDimensions-examplePageElement-new', {resizeDimensions: {left: 15, top: 250}});
+  xit('should do a save element with custom dimensions the new way', async () => {
+    await browser.imageCompare.saveElement(firstButton, 'resizeDimensions-firstButtonElement-new', {resizeDimensions: {left: 15, top: 250}});
   });
 });